Install: QMK: Method 1
Git operations (for example, as part of oh-my-posh, so even just pressing enter in the terminal) in the QMK repo are slow. I initially wondered if this was a problem caused by cloning the repo in MSYS and tried method 2. It turns out that the addition of the submodules is causing the slowness (whether there is a large number of files or that is the nature of submodules or something else, I'm not sure). So there is no performance difference between this commit and the next.

I decided to go with method 2 because, even though method 1 seems to be the more canonical way, method 2 has much better output during cloning (a single line that updates with progress vs extra lines for each progress update). I wanted to keep this method in history for later reference, if needed.