PowerShell: Remove section that keeps failing
```
No match was found for the specified search criteria for the provider 'NuGet'. The package provider requires 'PackageManagement' and 'Provider' tags. Please check if the specified package has the tags.
```

This succeeds in Windows PowerShell but fails in PS Core
I'm not sure why it was really needed in the first place
The history of the line shows some volatility
Other commits around it make it look like it was for oh-my-posh originally